Council ggplot2 theme

Council ggplot2 theme

council_theme(
  use_showtext = FALSE,
  size_header = 22,
  size_axis_title = 14,
  size_legend_title = 14,
  size_axis_text = 11,
  size_legend_text = 10,
  size_caption = 8,
  size_margin = 10
)

Arguments

use_showtext
logical, whether to use Council fonts.

size_header
header font size in px

size_axis_title
axis title font size in px

size_legend_title
legend title font size in px

size_axis_text
axis text font size in px

size_legend_text
legend test font size in px

size_caption
caption font size in px

size_margin
the margin size in pt

Value

a ggplot2::theme() object

Council fonts

If you wish to use Council fonts, you must have the following fonts installed:

HelveticaNeueLT Standard Light, Standard Condensed, and Standard Medium Condensed

Arial Narrow

Palatino Linotype

You must also have the showtext and sysfonts packages installed.

Font size suggestions

If you are creating plots for a presentation, you should increase all font sizes. Check the default font sizes in the presentation platform you are working in and adjust from there.

Theme modifications

If you wish to make changes to this theme template, such as removing the legend, adusting text positons, access the source code by typing councilR::council_theme in the console. You can then copy/paste the function into whatever document you are working on and make changes.

See also