Install councilR snippets

This function installs a set of R snippets that are helpful when working with Metropolitan Council projects in in RStudio. By default, the snippets are installed where RStudio will find them. If you haven't previously installed snippets to RStudio, these snippets will mask some of the built-in snippets that ship with RStudio.

snippets_install(install_path = NULL, update = TRUE)

Arguments

install_path
Where should the snippets be installed? If NULL, the snippets will install to a default path based on the current version of RStudio.

update
Should existing snippets be updated in place if there are any conflicts? Default is yes (TRUE). Otherwise, new snippets are appended to the end of the existing file, ensuring that you can recover your previous snippets by editing the snippets file.

Note

This function is adapted from js4shiny::snippets_install().

Updating Existing Snippets

If you already have snippets installed, you can you can have the installed snippets update the existing snippets in place with update = TRUE. Or you can append the new snippets to the existing snippets files with update = FALSE. This option is desirable if you want to make sure that no snippets are overwritten. The newer snippets will mask older snippets, but no data will be lost.

Examples

if (FALSE) {
snip_tmp <- tempfile("snippets")
dir.create(snip_tmp)
snippets_install(snip_tmp)
}